Practica Unidad 1: Algoritmos

+++ Practica 1 Algoritmos/Autobus.java
@@ -1,5 +1,5 @@
 public class Autobus{
-   public double Autobuz(){
+   public int Autobuz(){
        int pase=45;//datos en bolivianos
        double viaje=1.8;
        double totalMes=pase/viaje;
